Language.Javascript.JQuery

Description

Module for accessing minified jQuery code (http://jquery.com/). As an example:

import qualified Language.Javascript.JQuery as JQuery

main = do
    putStrLn $ "jQuery version " ++ show JQuery.version ++ " source:"
    putStrLn =<< readFile =<< JQuery.file

Documentation

version :: Version

The version of jQuery provided by this package. Not necessarily the version of this package, but the versions will match in the first three digits.

file :: IO FilePath

A local file containing the minified jQuery code for version.

url :: String

A remote URL of the jQuery sources for version. The URL does not have a protocol prefix, so users may need to prepend either "http:" or "https:" (both work). The URL currently uses the jQuery CDN links at http://code.jquery.com/. Alternative CDN links are listed at http://jquery.com/download/#using-jquery-with-a-cdn.